Refacto broadcasting functions
